requirements.txt
# Add your requirements here like:
# numpy
# scipy>=0.9
graphql-core==2.2.1
gql==0.1.0
cryptography==3.1
redis==3.5.3
eiffellib==1.1.0
pydantic==1.6
python-box==5.2.0
